Update tempConverter
I added a loop so that you don't have to run the program all over again for each temperature you have (and hopefully fixed all output typos). 
There is a small error where if you mistype a number from the menu where you choose either f->c or c->f it will tell you to make sure you "only enter a 1 or a 2" but then instead of letting you type a 1 or a 2 right away, it asks you if you want to run the program again. It's not that big of an issue, but it can be a cause for minor frustration on the users end.
idk i might come back to that later and fix it